Added GCMT_event class and db creation functions
GCMT_event class will be the master class that stores info
and methods for each event; it will allow for updates,
exports to various formats, etc.

The db creation functions are basically ones that automate
the fetching of data from globalcmt.org and processing it.

Currently, creating the databases out of the catalog files
are supported by the functions. Scripts have been written
to do the updating from quick CMTs but this hasn't been
properly abstracted/functionalized yet.